pyproject.toml
[project]
name = "daily-brief"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "langchain-groq>=1.1.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]

[project.scripts]
brief = "daily_brief.cli.main:dailybrief"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["daily_brief"]
